Expose term GET endpoint
