+++ server/package.json
@@ -16,7 +16,7 @@
   "dependencies": {
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.18.3",
-    "cheerio": "^1.0.0-rc.2",
+    "cheerio": "^1.0.0-rc.3",
     "concurrently": "^4.1.0",
     "express": "^4.16.4",
     "html-entities": "^1.2.1",
